refactor(menu): split the user interface into several files
