+++ src/include/TUI/helper_functions.h
@@ -1,0 +1,5 @@
+#ifndef TUI_HELPER_FUNCTIONS_H
+#define TUI_HELPER_FUNCTIONS_H
+void m_is_prime(void);
+void m_get_prime(void);
+#endif // TUI_HELPER_FUNCTIONS_H
