+++ src/include/TUI/rsa.h
@@ -1,0 +1,7 @@
+#ifndef TUI_RSA_H
+#define TUI_RSA_H
+
+void m_rsa(void);
+void m_rsa_encrypt(void);
+
+#endif // TUI_RSA_H
